refactor(destination-postgres): split client into separate pg and pglite files
Committed-By-Agent: claude
